chore: refactor to isolate async_query
