+++ src/lib.rs
@@ -115,6 +115,8 @@ pub mod utils;
 pub mod variant;
 
 #[cfg(feature = "async-query")]
+pub mod async_query;
+#[cfg(feature = "async-query")]
 pub mod query_sink;
 
 #[cfg(any(test, feature = "test"))]
